+++ CHANGELOG.md
@@ -8,3 +8,4 @@ Features:
 - Implement `EventTopicGenerateCode` for TypeScript projects and JSON events.
 - Implement `ProjectGetArtefactDestination` for npm packages.
 - Implement the `NpmService`.
+- Implement `ProjectBuildArtefact` for TypeScript packages.